refactor: move puddle update logic to shared
