+++ Content.Shared/Fluids/SharedPuddleSystem.cs
@@ -7,8 +7,13 @@ using Content.Shared.DragDrop;
 using Content.Shared.Examine;
 using Content.Shared.FixedPoint;
 using Content.Shared.Fluids.Components;
+using Content.Shared.Friction;
+using Content.Shared.Movement.Components;
 using Content.Shared.Movement.Events;
+using Content.Shared.Movement.Systems;
+using Content.Shared.Slippery;
 using Content.Shared.StepTrigger.Components;
+using Content.Shared.StepTrigger.Systems;
 using Robust.Shared.Containers;
 using Robust.Shared.Map;
 using Robust.Shared.Prototypes;
@@ -21,8 +26,11 @@ public abstract partial class SharedPuddleSystem : EntitySystem
     [Dependency] private readonly IGameTiming _timing = default!;
     [Dependency] private readonly IPrototypeManager _prototypeManager = default!;
     [Dependency] private readonly SharedAppearanceSystem _appearance = default!;
-    [Dependency] private readonly SharedSolutionContainerSystem _solutionContainerSystem = default!;
     [Dependency] private readonly SharedDoAfterSystem _doAfterSystem = default!;
+    [Dependency] private readonly SharedSolutionContainerSystem _solutionContainerSystem = default!;
+    [Dependency] private readonly SpeedModifierContactsSystem _speedModContacts = default!;
+    [Dependency] private readonly StepTriggerSystem _stepTrigger = default!;
+    [Dependency] private readonly TileFrictionController _tile = default!;
 
     private static readonly ProtoId<ReagentPrototype> Blood = "Blood";
     private static readonly ProtoId<ReagentPrototype> Slime = "Slime";
@@ -36,6 +44,10 @@ public abstract partial class SharedPuddleSystem : EntitySystem
     public const float LowThreshold = 0.3f;
 
     public const float MediumThreshold = 0.6f;
+
+    // Using local deletion queue instead of the standard queue so that we can easily "undelete" if a puddle
+    // loses & then gains reagents in a single tick.
+    private HashSet<EntityUid> _deletionQueue = [];
 
     public override void Initialize()
     {
@@ -55,11 +67,35 @@ public abstract partial class SharedPuddleSystem : EntitySystem
         InitializeSpillable();
     }
 
-    protected virtual void OnSolutionUpdate(Entity<PuddleComponent> entity, ref SolutionContainerChangedEvent args)
+    public override void Update(float frameTime)
+    {
+        base.Update(frameTime);
+
+        foreach (var ent in _deletionQueue)
+        {
+            Del(ent);
+        }
+
+        _deletionQueue.Clear();
+
+        TickEvaporation();
+    }
+
+    private void OnSolutionUpdate(Entity<PuddleComponent> entity, ref SolutionContainerChangedEvent args)
     {
         if (args.SolutionId != entity.Comp.SolutionName)
             return;
 
+        if (args.Solution.Volume <= 0)
+        {
+            _deletionQueue.Add(entity);
+            return;
+        }
+
+        _deletionQueue.Remove(entity);
+        UpdateSlip((entity, entity.Comp), args.Solution);
+        UpdateSlow(entity, args.Solution);
+        UpdateEvaporation(entity, args.Solution);
         UpdateAppearance((entity, entity.Comp));
     }
 
@@ -180,6 +216,115 @@ public abstract partial class SharedPuddleSystem : EntitySystem
 
         _appearance.SetData(ent, PuddleVisuals.CurrentVolume, volume.Float(), appearance);
         _appearance.SetData(ent, PuddleVisuals.SolutionColor, color, appearance);
+    }
+
+    private void UpdateSlip(Entity<PuddleComponent> entity, Solution solution)
+    {
+        if (!TryComp<StepTriggerComponent>(entity, out var comp))
+            return;
+
+        // Ensure we actually have the component
+        EnsureComp<TileFrictionModifierComponent>(entity);
+
+        // This is the base amount of reagent needed before a puddle can be considered slippery. Is defined based on
+        // the sprite threshold for a puddle larger than 5 pixels.
+        var smallPuddleThreshold = FixedPoint2.New(entity.Comp.OverflowVolume.Float() * LowThreshold);
+
+        // Stores how many units of slippery reagents a puddle has
+        var slipperyUnits = FixedPoint2.Zero;
+        // Stores how many units of super slippery reagents a puddle has
+        var superSlipperyUnits = FixedPoint2.Zero;
+
+        // These three values will be averaged later and all start at zero so the calculations work
+        // A cumulative weighted amount of minimum speed to slip values
+        var puddleFriction = FixedPoint2.Zero;
+        // A cumulative weighted amount of minimum speed to slip values
+        var slipStepTrigger = FixedPoint2.Zero;
+        // A cumulative weighted amount of launch multipliers from slippery reagents
+        var launchMult = FixedPoint2.Zero;
+        // A cumulative weighted amount of stun times from slippery reagents
+        var stunTimer = TimeSpan.Zero;
+
+        // Check if the puddle is big enough to slip in to avoid doing unnecessary logic
+        if (solution.Volume <= smallPuddleThreshold)
+        {
+            _stepTrigger.SetActive(entity, false, comp);
+            _tile.SetModifier(entity, 1f);
+            return;
+        }
+
+        if (!TryComp<SlipperyComponent>(entity, out var slipComp))
+            return;
+
+        foreach (var (reagent, quantity) in solution.Contents)
+        {
+            var reagentProto = _prototypeManager.Index<ReagentPrototype>(reagent.Prototype);
+
+            // Calculate the minimum speed needed to slip in the puddle. Average the overall slip thresholds for all reagents
+            var deltaSlipTrigger = reagentProto.SlipData?.RequiredSlipSpeed ?? entity.Comp.DefaultSlippery;
+            slipStepTrigger += quantity * deltaSlipTrigger;
+
+            // Aggregate Friction based on quantity
+            puddleFriction += reagentProto.Friction * quantity;
+
+            if (reagentProto.SlipData == null)
+                continue;
+
+            slipperyUnits += quantity;
+            // Aggregate launch speed based on quantity
+            launchMult += reagentProto.SlipData.LaunchForwardsMultiplier * quantity;
+            // Aggregate stun times based on quantity
+            stunTimer += reagentProto.SlipData.ParalyzeTime * (float)quantity;
+
+            if (reagentProto.SlipData.SuperSlippery)
+                superSlipperyUnits += quantity;
+        }
+
+        // Turn on the step trigger if it's slippery
+        _stepTrigger.SetActive(entity, slipperyUnits > smallPuddleThreshold, comp);
+
+        // This is based of the total volume and not just the slippery volume because there is a default
+        // slippery for all reagents even if they aren't technically slippery.
+        slipComp.SlipData.RequiredSlipSpeed = (float)(slipStepTrigger / solution.Volume);
+        _stepTrigger.SetRequiredTriggerSpeed(entity, slipComp.SlipData.RequiredSlipSpeed);
+
+        // Divide these both by only total amount of slippery reagents.
+        // A puddle with 10 units of lube vs a puddle with 10 of lube and 20 catchup should stun and launch forward the same amount.
+        if (slipperyUnits > 0)
+        {
+            slipComp.SlipData.LaunchForwardsMultiplier = (float)(launchMult/slipperyUnits);
+            slipComp.SlipData.ParalyzeTime = (stunTimer/(float)slipperyUnits);
+        }
+
+        // Only make it super slippery if there is enough super slippery units for its own puddle
+        slipComp.SlipData.SuperSlippery = superSlipperyUnits >= smallPuddleThreshold;
+
+        // Lower tile friction based on how slippery it is, lets items slide across a puddle of lube
+        slipComp.SlipData.SlipFriction = (float)(puddleFriction/solution.Volume);
+        _tile.SetModifier(entity, slipComp.SlipData.SlipFriction);
+
+        Dirty(entity, slipComp);
+    }
+
+    private void UpdateSlow(EntityUid uid, Solution solution)
+    {
+        var maxViscosity = 0f;
+        foreach (var (reagent, _) in solution.Contents)
+        {
+            var reagentProto = _prototypeManager.Index<ReagentPrototype>(reagent.Prototype);
+            maxViscosity = Math.Max(maxViscosity, reagentProto.Viscosity);
+        }
+
+        if (maxViscosity > 0)
+        {
+            var comp = EnsureComp<SpeedModifierContactsComponent>(uid);
+            var speed = 1 - maxViscosity;
+            _speedModContacts.ChangeSpeedModifiers(uid, speed, comp);
+        }
+        else
+        {
+            RemComp<SpeedModifierContactsComponent>(uid);
+        }
     }
 
     public void DoTileReactions(TileRef tileRef, Solution solution)
